Change #1: Cartography
- Worked a little on the “classes” though decided to hold off on the
entities until after completing the world map on account of incomplete
understanding of inheritance. Currently facing issues with constructors
and multi-level virtual functions.
- Began work of WorldMap class, as well as text files representing the
various potential layouts. These maps are to be loaded in via 2D
arrays, and different text files correspond to different maps (forest,
desert, ocean…), with the numbers in each “tile” (rows/columns on the
text file) corresponding to a different type of tile. For now, this is
0 (ground), 1 (tree), 2 (water), and 9 (“player”/human), leaving 3~8
open. If more are needed, symbols and letters can be used.
-  Similarly, began work of maps themselves. Currently 8x8 grids,
though could be expanded to 16x16 later on. None are complete as of now.
- Realized I misspelled Testing Grounds. Testing Grouds it is! That’s
probably going to make it into the project name, somehow, knowing me…

Current Goals:

- Finish WorldMap class and begin fleshing out “maps,” and visual
representation of this map in a grid
- Work on BaseAI and Animal—rename to entity?—class, as well as derived
classes Mammal/Bacteria, and, by extension, derived classes Human and
Cat (test entities for now)
- Plan “game” itself.
- Get some more sleep.