Change #8.0: Lifeforms
- Created Entity class, as well as base class EHuman. Very much based
on Tile class, though will soon have some more functions to
differentiate.
- Removed derived tile THuman (just after finishing it, too…it had a
working inventory and everything) in order to replace it with EHuman!
- Created functional inventory system (in THuman), which is currently
not working with the switch to EHuman. Fixing this is a matter of
getting Entity class also interfacing with the HUD.
- Created text files for each entity file, loaded into an 2D entity
array (similarly to Tiles). This allows for different load-ins and
overlaying of the two to create Entities on top of tiles.
 - Shifted clicked element data in HUD to display on the left side. It
looks much neater now!